refactor(#51) - develop 충돌 및 오류 수정
