+++ src/test/java/com/flab/infrun/lecture/application/LectureCommandProcessorTest.java
@@ -1,15 +1,9 @@
 package com.flab.infrun.lecture.application;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-import com.flab.infrun.lecture.application.exception.DuplicateLectureFileNameException;
 import com.flab.infrun.lecture.domain.StubLectureRepository;
 import com.flab.infrun.member.domain.StubMemberRepository;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockMultipartFile;
 
 
@@ -26,25 +20,8 @@ class LectureCommandProcessorTest {
             lectureRepository,
             null,
             null,
-            memberRepository,
             null
         );
-    }
-
-    @Test
-    @DisplayName("강의 파일 이름 중복 불가능 exception 테스트")
-    void validateLectureFile() {
-
-        MockMultipartFile multipartFile1 = setupMultiPartTextFile("fileDup1", "test.txt",
-            "test file");
-        MockMultipartFile multipartFile2 = setupMultiPartTextFile("fileDup1", "test1.txt",
-            "test file2");
-        MockMultipartFile multipartFile3 = setupMultiPartTextFile("fileDup2", "test1.txt",
-            "test file2");
-
-        assertThatThrownBy(() -> processor.validateLectureFile(
-            List.of(multipartFile1, multipartFile2, multipartFile3))).isInstanceOf(
-            DuplicateLectureFileNameException.class);
     }
 
     private MockMultipartFile setupMultiPartTextFile(String fileName, String originFileName,
