+++ src/test/java/com/flab/infrun/lecture/domain/StubLectureRepository.java
@@ -1,6 +1,7 @@
 package com.flab.infrun.lecture.domain;
 
 import com.flab.infrun.lecture.domain.repository.LectureRepository;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
@@ -20,5 +21,10 @@ public final class StubLectureRepository implements LectureRepository {
     @Override
     public Optional<Lecture> findById(Long id) {
         return Optional.ofNullable(persistence.get(id));
+    }
+
+    @Override
+    public List<Lecture> findAllByIdIn(List<Long> lectureIds) {
+        return null;
     }
 }
