+++ src/test/java/com/flab/infrun/utils/persistence/StubMemberRepository.java
@@ -36,4 +36,9 @@ final class StubMemberRepository implements MemberRepository {
             .filter(member -> Objects.equals(member.getEmail(), email))
             .findAny();
     }
+
+    @Override
+    public Member findById(Long id) {
+        return null;
+    }
 }
